seperate out types and interfaces
